Added -o, --vars-optional args (variables no longer required)
When a file is fed to templater.sh to replace variables within the
template but the template has no variables to replace, the expected
behavior would be to print out the file as is. I've added this new
argument to not change existing behavior but allow for the option to no
longer require variables within the file. This is handy if you have a
folder of files that may or may not need variable substitution.